Update embyupdate.py
Now pulling the master release API instead of just the /latest API. That exposed the beta versions. Also created a new "installbeta" Boolean user can change to install either the beta version or the stable version.  Added logic to search for the most recent of either beta or stable and install it.